refactor: refactor EditableCell : make it reuseable. refactor all table: use shared EditableCell.
