--- src/components/views/DatabaseView/ConsignmentTable/type/cellType.ts
@@ -1,9 +1,0 @@
-import { Table } from "@tanstack/react-table";
-import { Consignment } from "./dataType";
-
-export type CellType = {
-  getValue: () => string | number;
-  row: { index: number };
-  column: { id: string };
-  table: Table<Consignment>;
-};
