+++ src/components/views/DatabaseView/CustomTable/Columns.ts
@@ -1,6 +1,6 @@
 import { createColumnHelper } from "@tanstack/react-table";
 import { Custom } from "./type/dataType";
-import EditableCell from "./EditableCell";
+import EditableCell from "../shared/EditableCell";
 
 const columnHelper = createColumnHelper<Custom>();
 
@@ -8,27 +8,27 @@ const columns = [
   columnHelper.accessor("storeName", {
     cell: EditableCell,
     header: "店名",
-    enableColumnFilter: true, 
+    enableColumnFilter: true,
   }),
   columnHelper.accessor("contactPerson", {
     cell: EditableCell,
     header: "聯絡人",
-    enableColumnFilter: true, 
+    enableColumnFilter: true,
   }),
   columnHelper.accessor("landlinePhone", {
     cell: EditableCell,
     header: "市話",
-    enableColumnFilter: false, 
+    enableColumnFilter: false,
   }),
   columnHelper.accessor("mobilePhone", {
     cell: EditableCell,
     header: "手機",
-    enableColumnFilter: false, 
+    enableColumnFilter: false,
   }),
   columnHelper.accessor("address", {
     cell: EditableCell,
     header: "地址",
-    enableColumnFilter: false, 
+    enableColumnFilter: false,
   }),
 ];
