--- src/components/views/DatabaseView/ExampleTable/type/cellType.ts
@@ -1,9 +1,0 @@
-import { Table } from "@tanstack/react-table";
-import { Person } from "./dataType";
-
-export type CellType = {
-  getValue: () => string;
-  row: { index: number };
-  column: { id: string };
-  table: Table<Person>;
-};
