--- src/components/views/DatabaseView/InventoryTable/type/cellType.ts
@@ -1,9 +1,0 @@
-import { Table } from "@tanstack/react-table";
-import { Inventory } from "./dataType";
-
-export type CellType = {
-  getValue: () => string;
-  row: { index: number };
-  column: { id: string };
-  table: Table<Inventory>;
-};
