+++ src/components/views/DatabaseView/shared/index.ts
@@ -1,0 +1,2 @@
+// 共用組件的統一導出
+export { default as EditableCell } from "./EditableCell/index";
